moved scoring types to separate modules
